color: Use repr transparent instead of packed
Color is a single field struct with a 4 byte value. It is unnecessary to
pack it, because it will always be aligned. Mark it as transparent so
the compiler can treat it as a plain u32.

Add a test to show that arrays of colors contain no extra padding.

Use core::fmt so Debug is available in all environments.

Signed-off-by: Tim Crawford <tcrawford@system76.com>